Refactor Spells update page to use resource update-page.
