--- app/javascript/spells/store/deleteSpell/index.js
@@ -1,13 +1,0 @@
-import deleteEndpoint from '../../../api/resources/delete';
-
-const endpoint = deleteEndpoint({ resourceName: 'spell' });
-
-export default endpoint;
-
-export const {
-  actions,
-  hooks,
-  namespace,
-  reducer,
-  request,
-} = endpoint;
