+++ app/javascript/spells/store/index.js
@@ -1,18 +1,13 @@
 import { combineReducers } from 'redux';
 
-import { reducer as deleteSpell } from './deleteSpell';
-import { reducer as updateFindSpell } from './updateFindSpell';
-import { reducer as updateSpellForm } from './updateSpellForm';
-
 import { reducer as create } from '../pages/create-page';
 import { reducer as index } from '../pages/index-page';
 import { reducer as show } from '../pages/show-page';
+import { reducer as update } from '../pages/update-page';
 
 export default combineReducers({
   create,
-  deleteSpell,
   index,
   show,
-  updateFindSpell,
-  updateSpellForm,
+  update,
 });
